+++ dlib/matrix/lapack/geqrf.h
@@ -145,7 +145,7 @@ namespace dlib
                 return info;
 
             if (work.size() < work_size)
-                work.set_size(work_size, 1);
+                work.set_size(static_cast<long>(work_size), 1);
 
             // compute the actual decomposition 
             info = binding::geqrf(a.nr(), a.nc(), &a(0,0), a.nr(),
